requirement.txt
streamlit
langchain
langchain-openai
chromadb
python-dotenv
openai
